+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,5 +1,6 @@
 # unreleased
 - SearchListItem component: new `countLabel` prop that will overwrite the `item.count` value.
+- Add new component `<List />` for displaying interactive list items.
 
 # 3.1.0
 - Added support for a countLabel prop on SearchListItem to allow custom counts.